+++ accounts/urls.py
@@ -1,7 +1,7 @@
 from django.conf.urls import url
 from rest_framework.urlpatterns import format_suffix_patterns
 
-from socialnetworkauth import views
+from accounts import views
 
 
 urlpatterns = [
