+++ socialnetwork/urls.py
@@ -18,5 +18,5 @@ from django.contrib import admin
 
 urlpatterns = [
     url(r'^admin/', admin.site.urls),
-    url(r'^api/v1/', include('socialnetworkauth.urls')),
+    url(r'^api/v1/', include('accounts.urls')),
 ]
